refactor: share the multi-select filter checklist across the toolbars
